BLS12-381 precompiles for SDK refactoring
