+++ engine-precompiles/src/bls12_381/g1_add.rs
@@ -1,4 +1,3 @@
-use super::G1_INPUT_ITEM_LENGTH;
 use crate::prelude::types::{make_address, Address, EthGas};
 use crate::prelude::Borrowed;
 use crate::{EvmPrecompileResult, Precompile, PrecompileOutput, Vec};
@@ -18,37 +17,14 @@ impl BlsG1Add {
 
     #[cfg(feature = "std")]
     fn execute(input: &[u8]) -> Result<Vec<u8>, ExitError> {
-        use super::standalone::g1;
-        use blst::{
-            blst_p1, blst_p1_add_or_double_affine, blst_p1_affine, blst_p1_from_affine,
-            blst_p1_to_affine,
-        };
-
-        // NB: There is no subgroup check for the G1 addition precompile.
-        //
-        // We set the subgroup checks here to `false`
-        let a_aff = &g1::extract_g1_input(&input[..G1_INPUT_ITEM_LENGTH], false)?;
-        let b_aff = &g1::extract_g1_input(&input[G1_INPUT_ITEM_LENGTH..], false)?;
-
-        let mut b = blst_p1::default();
-        // SAFETY: b and b_aff are blst values.
-        unsafe { blst_p1_from_affine(&mut b, b_aff) };
-
-        let mut p = blst_p1::default();
-        // SAFETY: p, b and a_aff are blst values.
-        unsafe { blst_p1_add_or_double_affine(&mut p, &b, a_aff) };
-
-        let mut p_aff = blst_p1_affine::default();
-        // SAFETY: p_aff and p are blst values.
-        unsafe { blst_p1_to_affine(&mut p_aff, &p) };
-
-        Ok(g1::encode_g1_point(&p_aff))
+        aurora_engine_sdk::bls12_381::g1_add(input)
+            .map_err(|e| ExitError::Other(Borrowed(e.as_ref())))
     }
 
     #[cfg(not(feature = "std"))]
     #[allow(clippy::range_plus_one)]
     fn execute(input: &[u8]) -> Result<Vec<u8>, ExitError> {
-        use super::{extract_g1, padding_g1_result, FP_LENGTH};
+        use super::{extract_g1, padding_g1_result, FP_LENGTH, G1_INPUT_ITEM_LENGTH};
 
         let (p0_x, p0_y) = extract_g1(&input[..G1_INPUT_ITEM_LENGTH])?;
         let (p1_x, p1_y) = extract_g1(&input[G1_INPUT_ITEM_LENGTH..])?;
