+++ engine-precompiles/src/bls12_381/g1_msm.rs
@@ -1,4 +1,4 @@
-use super::{msm_required_gas, G1_INPUT_ITEM_LENGTH, SCALAR_LENGTH};
+use super::msm_required_gas;
 use crate::prelude::types::{make_address, Address, EthGas};
 use crate::prelude::{Borrowed, Vec};
 use crate::{EvmPrecompileResult, Precompile, PrecompileOutput};
@@ -29,58 +29,15 @@ impl BlsG1Msm {
 
     #[cfg(feature = "std")]
     fn execute(input: &[u8]) -> Result<Vec<u8>, ExitError> {
-        use super::standalone::{extract_scalar_input, g1, NBITS};
-        use blst::{blst_p1, blst_p1_affine, blst_p1_from_affine, blst_p1_to_affine, p1_affines};
-
-        let k = input.len() / INPUT_LENGTH;
-        let mut g1_points: Vec<blst_p1> = Vec::with_capacity(k);
-        let mut scalars: Vec<u8> = Vec::with_capacity(k * SCALAR_LENGTH);
-        for i in 0..k {
-            let slice = &input[i * INPUT_LENGTH..i * INPUT_LENGTH + G1_INPUT_ITEM_LENGTH];
-
-            // BLST batch API for p1_affines blows up when you pass it a point at infinity, so we must
-            // filter points at infinity (and their corresponding scalars) from the input.
-            if slice.iter().all(|i| *i == 0) {
-                continue;
-            }
-
-            // NB: Scalar multiplications, MSMs and pairings MUST perform a subgroup check.
-            //
-            // So we set the subgroup_check flag to `true`
-            let p0_aff = &g1::extract_g1_input(slice, true)?;
-
-            let mut p0 = blst_p1::default();
-            // SAFETY: p0 and p0_aff are blst values.
-            unsafe { blst_p1_from_affine(&mut p0, p0_aff) };
-            g1_points.push(p0);
-
-            scalars.extend_from_slice(
-                &extract_scalar_input(
-                    &input[i * INPUT_LENGTH + G1_INPUT_ITEM_LENGTH
-                        ..i * INPUT_LENGTH + G1_INPUT_ITEM_LENGTH + SCALAR_LENGTH],
-                )?
-                .b,
-            );
-        }
-
-        // return infinity point if all points are infinity
-        if g1_points.is_empty() {
-            return Ok([0; 128].into());
-        }
-
-        let points = p1_affines::from(&g1_points);
-        let multiexp = points.mult(&scalars, NBITS);
-
-        let mut multiexp_aff = blst_p1_affine::default();
-        // SAFETY: multiexp_aff and multiexp are blst values.
-        unsafe { blst_p1_to_affine(&mut multiexp_aff, &multiexp) };
-
-        Ok(g1::encode_g1_point(&multiexp_aff))
+        aurora_engine_sdk::bls12_381::g1_msm(input)
+            .map_err(|e| ExitError::Other(Borrowed(e.as_ref())))
     }
 
     #[cfg(not(feature = "std"))]
     fn execute(input: &[u8]) -> Result<Vec<u8>, ExitError> {
-        use super::{extract_g1, padding_g1_result, FP_LENGTH};
+        use super::{
+            extract_g1, padding_g1_result, FP_LENGTH, G1_INPUT_ITEM_LENGTH, SCALAR_LENGTH,
+        };
 
         let k = input.len() / INPUT_LENGTH;
         let mut g1_input = crate::vec![0u8; k * (2 * FP_LENGTH + SCALAR_LENGTH)];
