+++ engine-precompiles/src/bls12_381/map_fp2_to_g2.rs
@@ -15,24 +15,8 @@ impl BlsMapFp2ToG2 {
 
     #[cfg(feature = "std")]
     fn execute(input: &[u8]) -> Result<Vec<u8>, ExitError> {
-        use super::standalone::g2;
-        use super::{remove_padding, PADDED_FP_LENGTH};
-        use blst::{blst_map_to_g2, blst_p2, blst_p2_affine, blst_p2_to_affine};
-
-        let input_p0_x = remove_padding(&input[..PADDED_FP_LENGTH])?;
-        let input_p0_y = remove_padding(&input[PADDED_FP_LENGTH..PADDED_FP2_LENGTH])?;
-        let fp2 = g2::check_canonical_fp2(input_p0_x, input_p0_y)?;
-
-        let mut p = blst_p2::default();
-        // SAFETY: p and fp2 are blst values.
-        // third argument is unused if null.
-        unsafe { blst_map_to_g2(&mut p, &fp2, core::ptr::null()) };
-
-        let mut p_aff = blst_p2_affine::default();
-        // SAFETY: p_aff and p are blst values.
-        unsafe { blst_p2_to_affine(&mut p_aff, &p) };
-
-        Ok(g2::encode_g2_point(&p_aff))
+        aurora_engine_sdk::bls12_381::map_fp2_to_g12(input)
+            .map_err(|e| ExitError::Other(Borrowed(e.as_ref())))
     }
 
     #[cfg(not(feature = "std"))]
