+++ engine-sdk/Cargo.toml
@@ -15,6 +15,7 @@ publish = true
 [dependencies]
 aurora-engine-types.workspace = true
 base64.workspace = true
+blst = { workspace = true, optional = true }
 libsecp256k1 = { workspace = true, features = ["static-context", "hmac"] }
 sha2.workspace = true
 sha3.workspace = true
@@ -26,7 +27,7 @@ hex.workspace = true
 workspace = true
 
 [features]
-std = ["aurora-engine-types/std", "libsecp256k1/std", "sha3/std", "sha2/std", "base64/std"]
+std = ["aurora-engine-types/std", "libsecp256k1/std", "sha3/std", "sha2/std", "base64/std", "blst"]
 contract = []
 log = []
 all-promise-actions = []
